% Generated by roxygen2 (4.0.2): do not edit by hand
\name{authenticate}
\alias{authenticate}
\title{Authenticate with twitter api}
\usage{
authenticate()
}
\arguments{
\item{none}{}
}
\value{
oauth signature object to be used in future requests
}
\description{
This function authenticates with the twitter api using the `httr` package.
}
\keyword{misc}

