% Generated by roxygen2 (4.0.2): do not edit by hand
\name{embedTweet}
\alias{embedTweet}
\title{Embed a tweet into your R markdown}
\usage{
embedTweet(id = "506900131086880768")
}
\arguments{
\item{id}{}
}
\value{
string html object with a tweet's text, username, and user avatar
}
\description{
This function returns a tweet's text, the username, and the user's avatar wrapped in html code
}
\details{
This function solves a simple problem. You want to embed a tweet into your R markdown to impress
your professor and your peers?
You can do that by simply calling the functions from this package.
}
\note{
tweet id can be obtained from the link to a tweet. Example: "https://twitter.com/STAT545/status/506900131086880768"
}
\examples{
embedTweet(id = "506900131086880768")
}
\keyword{misc}

