% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{funWithTwitter}
\alias{funWithTwitter}
\alias{funWithTwitter-package}
\title{R Package: funWithTwitter}
\description{
This package can get you a tweet using its ID and embed the tweet into R markdown.
}
\details{
Single paragraph with more detail on the package.
This can span several lines.
}

