% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getTweet}
\alias{getTweet}
\title{Get a tweet from its ID}
\usage{
getTweet(id = "506900131086880768")
}
\arguments{
\item{id}{}
}
\value{
character string with the tweet text, and user name
}
\description{
This function returns a tweet's text, and the username
}
\details{
This function solves a simple problem. You want to embed a tweet into your R markdown to impress
your professor and your peers?
You can do that by simply calling the functions from this package.
}
\note{
tweet id can be obtained from the link to a tweet. Example: "https://twitter.com/STAT545/status/506900131086880768"
}
\examples{
getTweet(id = "506900131086880768")
}
\keyword{misc}

