% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getUserTimeline}
\alias{getUserTimeline}
\title{Get the timeline of a twitter user}
\usage{
getUserTimeline(screen_name = "STAT545", count = 1,
  exclude_replies = TRUE, include_rts = TRUE)
}
\arguments{
\item{screen_name}{Twitter username}

\item{count}{The number of tweets to return}

\item{exclude_replies}{if TRUE, do not include replies in results}

\item{include_rts}{if TRUE, include retweets in results}
}
\value{
vector of character ids of tweets
}
\description{
This function returns tweet ids of a twitter user's timeline in chronological order
}
\details{
This is a function to return tweet ids from a twitter user's timeline starting from the most recent tweet.
The results can be used as input for `getTweet()` or `embedTweet()` functions.
}
\note{
Adheres to the limitations imposed by the twitter api. For example, count limit is 200.
}
\examples{
getUserTimeline(screen_name = "STAT545", count = 5, exclude_replies = FALSE, include_rts = FALSE)
}
\keyword{misc}

