% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getTweet}
\alias{getTweet}
\title{Get a tweet from its ID}
\usage{
getTweet(id = "506900131086880768")
}
\arguments{
\item{id}{a tweet's id}
}
\value{
character string with the tweet text, and user name
}
\description{
This function returns a tweet's text, and the username
}
\details{
This is a function to check a tweet from its id.
Enter an id and the function will return the usename and text of the tweet.
}
\note{
Tweet id can be obtained from the link to a tweet. Example: "https://twitter.com/STAT545/status/506900131086880768"
}
\examples{
getTweet(id = "506900131086880768")
}
\keyword{misc}

